enhancement/use-stdev-for-candidate-selection (#4)
This changes the cslb algorithm to look for candidates only if the standard deviation from the mean weight is greater than the new disparity tolerance metric (percentage of mean weight, default 20%).  This allows the algorithm to calm down when the load disparity is very low, preventing situations where the algorithm would constantly shuffle the same workload between a small number of hosts with similar weights.  So far it seems to be working well on my cluster.

Co-authored-by: padthaitofuhot <padthaitofuhot@users.noreply.github.com>
Reviewed-on: https://forgejo.praxis.red/siv/pve-cslb/pulls/4